Add measurements to admin

+++ database/migrations/1724355195589_CreateUserLikedRecipeTable.ts
@@ -4,7 +4,7 @@ export async function up(db: Kysely<any>): Promise<void> {
 	await sql`
         create table user_liked_recipe (
            user_id bigint references users(id) on delete cascade,
-           recipe_id bigint references recipes(id) on delete cascade,
+           recipe_id integer references recipes(id) on delete cascade,
            primary key (user_id, recipe_id)
         )
     `.execute(db);
